Ignore Userspace and Local Home
When registering worlds, new worlds will only be added when the world count is greater than 2 so the userspace and local home are ignored. This somewhat falls apart if the local home crashes, but it's expected that the local home and userspace will always be present and I do not want to account for them crashing.